requirements.txt
# Django
Django==5.0.1

# Environnement
python-decouple==3.8

# API
requests==2.31.0

# Images
Pillow==10.2.0

# Production
gunicorn==21.2.0
psycopg2-binary==2.9.9
whitenoise==6.6.0
dj-database-url==2.1.0

# Développement (optionnel)
python-dotenv==1.0.0